rename files in lib
