+++ lib/cmd-start.js
