+++ lib/graph-assets.js
