+++ lib/graph-document.js
