+++ lib/graph-manifest.js
