+++ lib/graph-script.js
