+++ lib/graph-service-worker.js
