+++ lib/graph-style.js
